% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{interact}
\alias{interact}
\title{Interaction}
\usage{
interact(vars)
}
\arguments{
\item{...}{}
}
\description{
...
}
\section{Integration into ggplot2}{

You can simply add to a plot created by \code{run2elev} by using the usual \code{ggplot2}-syntax.
}

\examples{
\dontrun{
update_yaml("voteR",overwrite = FALSE)
}
}
