% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{reg}
\alias{reg}
\title{Run regression}
\usage{
reg(method = "lm", depvar = "1", covars = list(c("cyl", "disp"), c("drat",
  "vs", "am")), data, stepwise = FALSE, stargazer = FALSE, ...)
}
\arguments{
\item{...}{}

\item{...}{Other parameters passed into the regression function}
}
\description{
...
}
\section{Integration into ggplot2}{

You can simply add to a plot created by \code{run2elev} by using the usual \code{ggplot2}-syntax.
}

\examples{
\dontrun{
reg(method = "lm",
    depvar = "mpg",
    covars = list(c("cyl","disp"),
                  c("drat","vs","am")),
    data = "mtcars",
    stepwise = T,
    stargazer = T)
}
}
