% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_yaml.R
\name{update_yaml}
\alias{update_yaml}
\title{Update Package YAML file}
\usage{
update_yaml(mypkg, overwrite = FALSE)
}
\arguments{
\item{mypkg}{....}

\item{overwrite}{....}
}
\description{
...
}
\section{Integration into ggplot2}{

You can simply add to a plot created by \code{run2elev} by using the usual \code{ggplot2}-syntax.
}

\examples{
\dontrun{
update_yaml("voteR",overwrite = FALSE)
}
}
